Default Transplanting Report Test (#110)
__Pull Request Description__

Addresses #11 

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Co-authored-by: AdacusG <greenead@dickionson.edu>
Co-authored-by: AzizbekMuminoff <40713295+AzizbekMuminoff@users.noreply.github.com>
Co-authored-by: hoangphung-11 <phungthehoang11@gmail.com>