Created Tests for Issue#10 (#114)
__Pull Request Description__

Address #10 
Tested the default values in the SeedingReport
 - page contains the header "Seeding Report"

- page contains a section labeled "Set Dates"

- button has the label "Generate Report" and is enabled.

- default start date is the first day of the current year.

- default end date is the current date.

- the remainder of the report is not visible or does not exist.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Co-authored-by: btaniemie <lennminh@gmail.com>
Co-authored-by: Minh Le <81462198+btaniemie@users.noreply.github.com>