Test Seeding Input Type Defaults (#113)
__Pull Request Description__

Includes 5 tests to check the default values of the seeding type option
on the seeding input page. Closes #13.

---
__Licensing Certification__

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request __I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)__ for its
contents.

---------

Signed-off-by: Lukas Hamm <hamml@dickinson.edu>
Co-authored-by: pediconj <pediconj@dickinson.edu>
Co-authored-by: bussm <bussm@dickinson.edu>